Alteração para uso de arquivos de contexto para resolver erros contantes de busca de IDs

+++ app/(tabs)/patients/PacienteDetails.tsx
@@ -1,43 +1,44 @@
 // app/(tabs)/patients/PacienteDetails.tsx
 
-import { useRouter, useLocalSearchParams } from 'expo-router';
+import { useRouter } from 'expo-router';
 import React, { useEffect, useState } from 'react';
 import { View, Text, TouchableOpacity, Alert, ScrollView, SafeAreaView } from 'react-native';
 
 import LoadingOverlay from '../../../components/LoadingOverlay';
-import { Patient, Attendance, Vaccination } from '../../../powersync/AppSchema';
+import { Attendance, Vaccination } from '../../../powersync/AppSchema';
 import { useSystem } from '../../../powersync/PowerSync';
 import styles from '../../styles/PacienteDetailsStyles';
+import { usePatient } from '../../context/PatientContext';
+import { useDoctor } from '../../context/DoctorContext';
 
 const PacienteDetails = () => {
   const router = useRouter();
-  const { patient } = useLocalSearchParams();
-  const parsedPatient: Patient | null = patient
-    ? JSON.parse(decodeURIComponent(patient as string))
-    : null;
-
   const { db, supabaseConnector } = useSystem();
   const [loading, setLoading] = useState(true);
   const [attendance, setAttendance] = useState<Attendance | null>(null);
   const [hasVaccinations, setHasVaccinations] = useState<boolean>(false);
 
+  // Obtendo o paciente e o médico do contexto
+  const { selectedPatient } = usePatient();
+  const { doctorId } = useDoctor();
+
   useEffect(() => {
     const fetchPatientDetails = async () => {
-      if (!parsedPatient) {
+      if (!selectedPatient) {
         Alert.alert('Erro', 'Paciente não encontrado.');
         router.replace('/(tabs)/home/');
         setLoading(false);
         return;
       }
-      await checkAttendance(parsedPatient.id);
-      await checkVaccinations(parsedPatient.id);
+      await checkAttendance(selectedPatient.id);
+      await checkVaccinations(selectedPatient.id);
     };
 
     fetchPatientDetails().catch((error) => {
       console.error('Erro ao buscar detalhes do paciente:', error);
       setLoading(false);
     });
-  }, []);
+  }, [selectedPatient]);
 
   const checkAttendance = async (patientId: string) => {
     try {
@@ -83,17 +84,16 @@ const PacienteDetails = () => {
   };
 
   const handleOpenConsulta = () => {
-    const encodedPatient = encodeURIComponent(JSON.stringify(parsedPatient));
-    if (parsedPatient?.id) {
+    if (selectedPatient?.id) {
       if (attendance) {
         // Se já existe um prontuário, redirecionar para a atualização do prontuário
         router.push(
-          `/attendences/UpdateAttendance?patient=${encodedPatient}&attendanceId=${attendance.id}` as unknown as `${string}:${string}`
+          `/attendences/UpdateAttendance?patientId=${selectedPatient.id}&attendanceId=${attendance.id}` as unknown as `${string}:${string}`
         );
       } else {
         // Se não existe um prontuário, redirecionar para o registro de um novo prontuário
         router.push(
-          `/attendences/RegisterAttendance?patientId=${parsedPatient.id}` as unknown as `${string}:${string}`
+          `/attendences/RegisterAttendance?patientId=${selectedPatient.id}` as unknown as `${string}:${string}`
         );
       }
     } else {
@@ -102,15 +102,21 @@ const PacienteDetails = () => {
   };
 
   const handleRegisterVaccination = () => {
-    const encodedPatient = encodeURIComponent(JSON.stringify(parsedPatient));
-    router.push(
-      `/vaccines/RegisterVaccination?patient=${encodedPatient}` as unknown as `${string}:${string}`
-    );
+    if (selectedPatient?.id) {
+      router.push(
+        `/vaccines/RegisterVaccination?patientId=${selectedPatient.id}` as unknown as `${string}:${string}`
+      );
+    } else {
+      Alert.alert('Erro', 'ID do paciente não encontrado.');
+    }
   };
 
   const handleViewVaccinationCard = () => {
-    const encodedPatient = encodeURIComponent(JSON.stringify(parsedPatient));
-    router.push(`/vaccines/?patient=${encodedPatient}` as unknown as `${string}:${string}`);
+    if (selectedPatient?.id) {
+      router.push(`/vaccines/?patientId=${selectedPatient.id}` as unknown as `${string}:${string}`);
+    } else {
+      Alert.alert('Erro', 'ID do paciente não encontrado.');
+    }
   };
 
   const handleDeletePaciente = async () => {
@@ -126,26 +132,28 @@ const PacienteDetails = () => {
             setLoading(true);
             try {
               // Deletar todos os dados relacionados
-              await supabaseConnector.client
-                .from('attendances')
-                .delete()
-                .eq('patient_id', parsedPatient!.id);
-              await supabaseConnector.client
-                .from('vaccinations')
-                .delete()
-                .eq('patient_id', parsedPatient!.id);
-              const { error: deletePatientError } = await supabaseConnector.client
-                .from('patients')
-                .delete()
-                .eq('cpf', parsedPatient!.cpf);
+              if (selectedPatient) {
+                await supabaseConnector.client
+                  .from('attendances')
+                  .delete()
+                  .eq('patient_id', selectedPatient.id);
+                await supabaseConnector.client
+                  .from('vaccinations')
+                  .delete()
+                  .eq('patient_id', selectedPatient.id);
+                const { error: deletePatientError } = await supabaseConnector.client
+                  .from('patients')
+                  .delete()
+                  .eq('cpf', selectedPatient.cpf);
 
-              if (deletePatientError) {
-                console.error('Erro ao excluir paciente do Supabase:', deletePatientError);
-                Alert.alert('Erro', 'Erro ao excluir o paciente do Supabase.');
-              } else {
-                await db.deleteFrom('patients').where('cpf', '=', parsedPatient!.cpf).execute();
-                Alert.alert('Sucesso', 'Paciente e prontuários excluídos com sucesso.');
-                router.replace('/(tabs)/home/');
+                if (deletePatientError) {
+                  console.error('Erro ao excluir paciente do Supabase:', deletePatientError);
+                  Alert.alert('Erro', 'Erro ao excluir o paciente do Supabase.');
+                } else {
+                  await db.deleteFrom('patients').where('cpf', '=', selectedPatient.cpf).execute();
+                  Alert.alert('Sucesso', 'Paciente e prontuários excluídos com sucesso.');
+                  router.replace('/(tabs)/home/');
+                }
               }
             } catch (error) {
               console.error('Erro ao excluir paciente:', error);
@@ -168,41 +176,28 @@ const PacienteDetails = () => {
       <ScrollView contentContainerStyle={{ padding: 16 }}>
         <Text style={styles.header}>Detalhes do Paciente</Text>
         <View style={styles.detailsContainer}>
-          <Text style={styles.detailItem}>Nome: {parsedPatient?.name}</Text>
+          <Text style={styles.detailItem}>Nome: {selectedPatient?.name}</Text>
           {/* ... other patient details ... */}
         </View>
         <View style={styles.buttonContainer}>
           <TouchableOpacity style={styles.buttonDelete} onPress={handleDeletePaciente}>
             <Text style={styles.buttonText}>DELETAR PACIENTE</Text>
           </TouchableOpacity>
-          <TouchableOpacity
-            style={styles.buttonConsulta}
-            onPress={handleOpenConsulta}
-            disabled={!!attendance}
-          >
+          <TouchableOpacity style={styles.buttonConsulta} onPress={handleOpenConsulta}>
             <Text style={styles.buttonText}>ABRIR CONSULTA</Text>
           </TouchableOpacity>
           <TouchableOpacity
             style={styles.buttonConsulta}
             onPress={handleOpenConsulta}
-            disabled={!attendance}
-          >
+            disabled={!attendance}>
             <Text style={styles.buttonText}>NOVA CONSULTA</Text>
           </TouchableOpacity>
         </View>
         <View style={styles.buttonContainer}>
-          <TouchableOpacity
-            style={styles.buttonVaccine}
-            onPress={handleRegisterVaccination}
-            disabled={hasVaccinations}
-          >
+          <TouchableOpacity style={styles.buttonVaccine} onPress={handleRegisterVaccination}>
             <Text style={styles.buttonText}>REGISTRAR VACINA</Text>
           </TouchableOpacity>
-          <TouchableOpacity
-            style={styles.buttonVaccine}
-            onPress={handleViewVaccinationCard}
-            disabled={!hasVaccinations}
-          >
+          <TouchableOpacity style={styles.buttonVaccine} onPress={handleViewVaccinationCard}>
             <Text style={styles.buttonText}>CADERNETA DE VACINAÇÃO</Text>
           </TouchableOpacity>
         </View>
